requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.23
botocore==1.38.23
celery==5.3.6
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
decouple==0.0.7
distro==1.9.0
dj-database-url==2.3.0
Django==5.2.4
django-allauth==65.4.1
django-apscheduler==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==4.1
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.15.2
dnspython==2.7.0
docutils==0.21.2
et_xmlfile==2.0.0
eventlet==0.40.1
fonttools==4.56.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jwt==1.3.1
kiwisolver==1.4.8
kombu==5.5.4
Levenshtein==0.27.1
mailersend==0.6.0
matplotlib==3.10.0
msgpack==1.1.1
numpy==2.2.2
openai==1.99.6
openpyxl==3.1.5
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
phonenumbers==9.0.5
pillow==11.1.0
platformshconfig==2.4.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.1
RapidFuzz==3.13.0
redis==6.2.0
reportlab==4.3.1
requests==2.32.3
s3transfer==0.13.0
setuptools==80.9.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
XlsxWriter==3.2.2
zope.event==5.1.1
zope.interface==7.2
